--- src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,29 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.dto.MenuGroupRequest;
-import kitchenpos.dto.MenuGroupResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class MenuGroupService {
-    private final MenuGroupDao menuGroupDao;
-
-    public MenuGroupService(final MenuGroupDao menuGroupDao) {
-        this.menuGroupDao = menuGroupDao;
-    }
-
-    @Transactional
-    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
-        MenuGroup menuGroup = new MenuGroup(menuGroupRequest.getName());
-        return MenuGroupResponse.of(menuGroupDao.save(menuGroup));
-    }
-
-    public List<MenuGroupResponse> list() {
-        return MenuGroupResponse.ofList(menuGroupDao.findAll());
-    }
-}
