+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,0 +1,29 @@
+package kitchenpos.menu.application;
+
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
+        MenuGroup menuGroup = new MenuGroup(menuGroupRequest.getName());
+        return MenuGroupResponse.of(menuGroupRepository.save(menuGroup));
+    }
+
+    public List<MenuGroupResponse> list() {
+        return MenuGroupResponse.ofList(menuGroupRepository.findAll());
+    }
+}
